pyproject.toml
[project]
name = "bosch-climate-solutions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
packages = [{ include = "backend", from = "src" }]

dependencies = [
    "ibm-watsonx-ai>=1.1.20",
    "ipykernel>=6.29.5",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.0.1",
    "streamlit>=1.39.0",
    "xlsxwriter>=3.2.0",
]
